requirements.txt
google-generativeai
weaviate-client
pytest
pytest-asyncio
ruff
watchdog
